+++ app/src/Generators/ControllerGenerator.php
@@ -4,6 +4,9 @@ namespace UmigameTech\Catapult\Generators;
 
 class ControllerGenerator extends Generator
 {
+    const METHOD_GET = 'get';
+    const METHOD_POST = 'post';
+
     public static function controllerName($entity)
     {
         return implode('', array_map(
@@ -11,6 +14,19 @@ class ControllerGenerator extends Generator
             explode('_', $entity['name'])
         )) . 'Controller';
     }
+
+    public static $actions = [
+        'index' => self::METHOD_GET,
+        'show' => self::METHOD_GET,
+        'new' => self::METHOD_GET,
+        'createConfirm' => self::METHOD_POST,
+        'create' => self::METHOD_POST,
+        'edit' => self::METHOD_GET,
+        'editConfirm' => self::METHOD_POST,
+        'update' => self::METHOD_POST,
+        'deleteConfirm' => self::METHOD_GET,
+        'delete' => self::METHOD_POST, // HTMLフォームからの送信だとDELETEメソッドは使えないので
+    ];
 
     public function generate($entity)
     {
